Intake | refactor actions based on form
Makes way for actions to be seperated based on form selection
