+++ client/app/intake/components/CancelButton.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import Button from '../../components/Button';
-import { toggleCancelModal } from '../redux/actions';
+import { toggleCancelModal } from '../actions/common';
 import { connect } from 'react-redux';
 import { bindActionCreators } from 'redux';
