+++ client/app/intake/constants.js
@@ -43,6 +43,12 @@ export const RAMP_INTAKE_STATES = {
   COMPLETED: 'COMPLETED'
 };
 
+export const ENDPOINT_NAMES = {
+  INTAKE: 'intake',
+  INTAKE_RAMP: 'intake-ramp',
+  INTAKE_RAMP_COMPLETE: 'intake-ramp-complete'
+};
+
 export const FORM_TYPES = {
   RAMP_ELECTION: { key: 'ramp_election',
     name: 'RAMP Opt-In Election Form' },
