+++ client/app/intake/pages/selectForm.jsx
@@ -4,7 +4,7 @@ import Button from '../../components/Button';
 import { Redirect } from 'react-router-dom';
 import { connect } from 'react-redux';
 import { bindActionCreators } from 'redux';
-import { setFormType } from '../redux/actions';
+import { setFormType } from '../actions/common';
 import { FORM_TYPES, PAGE_PATHS } from '../constants';
 import _ from 'lodash';
